+++ public/js/func.js
