+++ public/js/init.js
