+++ routes/about.js
@@ -1,0 +1,9 @@
+var express = require('express');
+var router = express.Router();
+
+
+router.get('/about', function(req, res) {
+  res.render('about', { title: 'About' });
+});
+
+module.exports = router;
