+++ views/about.ejs
@@ -1,0 +1,104 @@
+<!DOCTYPE html>
+<html>
+    <head>
+        <meta charset = "utf-8"></meta>
+        <meta name = "viewport" content = "width = device-width, initial-scale = 1">
+        <meta name = "description" content = "專屬於中正大學學生的排課系統——鳳梨課表">
+        <title>CCU Class Schedule</title>
+        <link rel = "stylesheet" href = "./css/tailwind.css">
+        <!-- Global site tag (gtag.js) - Google Analytics -->
+        <script async src = "https://www.googletagmanager.com/gtag/js?id=G-4V5KYBRM5L"></script>
+        <script>
+            window.dataLayer = window.dataLayer || [];
+            function gtag(){dataLayer.push(arguments);}
+            gtag('js', new Date());
+            gtag('config', 'G-4V5KYBRM5L');
+        </script>
+    </head>
+
+    <body>
+        <nav class = "bg-[#eacda3] text-black top-0 nav justify-center p-2 py-3 mt-0 w-full relative">
+            <div class = "nav justify-center">
+                <div class = 'flex justify-between items-center'>
+                    <div class = 'flex items-center w-full md:w-4/12'>
+                        <img src = './img/logo.png' class = 'h-10'>
+                    </div>
+                    <div class = 'right-0 hidden md:block'>
+                        <a class = "btn-head" href = "./index.html">回首頁</a>
+                        <a class = "btn-head mt-3" href = "./about.html">關於我們</a> 
+                    </div>
+                    <div class = 'block md:hidden'>
+                        <!-- navbar collapse icon -->
+                        <button type = 'button' class = 'py-2 px-2 rounded text-black' onclick = "collapse()">
+                            <svg xmlns = "http://www.w3.org/2000/svg" class = "h-6 w-6" fill = "none" viewBox = "0 0 24 24" stroke = "currentColor" stroke-width = "2">
+                                <path stroke-linecap = "round" stroke-linejoin = "round" d = "M4 6h16M4 10h16M4 14h16M4 18h16"/>
+                            </svg>
+                        </button>
+                    </div>
+                </div>
+            </div>
+        </nav>
+        <div class = 'w-full bg-[#eacda3] collapse' id = 'mobile_menu'>
+            <div class = 'grid grid-cols-1 w-10/12 mx-auto text-white'>
+                <a class = "btn-head" href = "./index.html">回首頁</a>
+                <a class = "btn-head" href = "./about.html">關於我們</a>
+            </div>
+        </div>
+        <div class = 'w-full mx-auto py-6'>
+            <div class = 'w-11/12 md:w-7/12 mx-auto my-3'>
+                <div class = 'text-center text-3xl font-semibold border-b-2 mb-3'>
+                    關於我們
+                </div>
+                <div class = "text-left text-xl mx-auto">
+                    我們是現在就讀於中正大學的學生，為了改善中正學生的排課選課體驗而建立了鳳梨課表團隊，<br>
+                    希望能透過我們所學和自身經驗打造一個更加舒適且符合中正學生所需的服務。<br>
+                    目前我們提供之功能如下<br><br>
+                    1. 手動輸入課程資訊(可保存於同裝置之同一瀏覽器)<br>
+                    2. 產生對應之課表<br>
+                    3. 下載課表圖檔(行動用戶則無此服務)<br><br>
+                    我們預計將在下個版本開放課程搜尋功能，有任何問題歡迎透過下方回饋表單告知。
+                </div>
+            </div>
+            <div class = 'w-11/12 md:w-7/12 mx-auto my-3'>
+                <div class = 'text-center text-3xl font-semibold border-b-2 mb-3'>
+                    版本更新紀錄
+                </div>
+                <div class = "text-left text-xl mx-auto">
+                    v a1.0.0 第一次公開測試版本開放<br>
+                    v a1.5.0 UI大規模重構與優化，準備加入搜尋功能<br>
+                    v a2.0.0 UI全面使用Tailwind改寫
+                </div>
+            </div>
+            <div class = 'w-11/12 md:w-7/12 mx-auto my-3'>
+                <div class = 'text-center text-3xl font-semibold border-b-2 mb-3'>
+                    成員介紹
+                </div>
+                <div class = 'grid grid-cols-1 md:grid-cols-4 mx-auto border'>
+                    <div class = 'col-span-1 mx-auto my-7 md:my-0'>
+                        <img class = 'inline-block rounded-full w-72 h-72 md:w-auto md:h-auto md:rounded-none' src = './img/unknown.jfif'>
+                    </div>
+                    <div class = 'col-span-1 text-xl flex items-center mx-auto'>
+                        創辦人 D1stance
+                    </div>
+                    <div class = 'col-span-1 mt-3 md:mt-0 md:col-span-2 text-xl text-center flex items-center mx-auto'>
+                        現在就讀於中正大學資訊工程學系一年級，<br>
+                        目前正在鑽研程式演算法競賽與網頁前後端工程。
+                    </div>
+                </div>
+            </div>
+            <div class = 'w-11/12 md:w-7/12 mx-auto my-3'>
+                <div class = 'text-center text-3xl font-semibold border-b-2 mb-3'>
+                    聯絡我們
+                </div>
+                <div class = "mx-auto">
+                    <iframe src = "https://docs.google.com/forms/d/e/1FAIpQLSfaZZ3jg5Zt00LhQg8gGSLf06aI1huyLYXULc2eeCoAYR_SMg/viewform?embedded=true" width = "100%" height = "1626" frameborder = "0" marginheight = "0" marginwidth = "0">載入中…</iframe>
+                </div>
+            </div>
+        </div>
+        <div class = "bg-[#eacda3] py-3 w-full flex items-center">
+            <div class = 'text-center mx-auto'>© Pineapple Class Schedule 2022</div>
+        </div>
+    </body>
+    <script src = "https://code.jquery.com/jquery-3.6.0.min.js" integrity = "sha256-/xUj+3OJU5yExlq6GSYGSHk7tPXikynS7ogEvDej/m4=" crossorigin = "anonymous"></script>
+    <script src = "./js/func.js"></script>
+</html>
