+++ views/index.ejs
@@ -2,10 +2,10 @@
 <html>
     <head>
         <meta charset = "utf-8"></meta>
-        <meta name = "viewport" content = "width = device-width, initial-scale = 1">
+        <meta name = "viewport" content = "width = device-width, initial-scale = 1.0">
         <meta name = "description" content = "專屬於中正大學學生的排課系統——鳳梨課表">
         <title>CCU Class Schedule</title>
-        <link rel = "stylesheet" href = "./css/tailwind.css">
+        <link rel = "stylesheet" href = "/css/tailwind.css">
         <!-- Global site tag (gtag.js) - Google Analytics -->
         <script async src = "https://www.googletagmanager.com/gtag/js?id=G-4V5KYBRM5L"></script>
         <script>
@@ -22,11 +22,11 @@
             <div class = "nav justify-center">
                 <div class = 'flex justify-between items-center'>
                     <div class = 'flex items-center w-full md:w-4/12'>
-                        <img src = './img/logo.png' class = 'h-10'>
+                        <img src = '/img/logo.png' class = 'h-10'>
                     </div>
                     <div class = 'right-0 hidden md:block'>
-                        <a class = "btn-head" href = "./index.html">回首頁</a>
-                        <a class = "btn-head mt-3" href = "./about.html">關於我們</a> 
+                        <a class = "btn-head" href = "index">回首頁</a>
+                        <a class = "btn-head mt-3" href = "about">關於我們</a> 
                     </div>
                     <div class = 'block md:hidden'>
                         <!-- navbar collapse icon -->
@@ -41,8 +41,8 @@
         </nav>
         <div class = 'w-full bg-[#eacda3] collapse' id = 'mobile_menu'>
             <div class = 'grid grid-cols-1 w-10/12 mx-auto text-white'>
-                <a class = "btn-head" href = "./index.html">回首頁</a>
-                <a class = "btn-head" href = "./about.html">關於我們</a>
+                <a class = "btn-head" href = "index">回首頁</a>
+                <a class = "btn-head" href = "about">關於我們</a>
             </div>
         </div>
         <div class = 'w-11/12 md:w-8/12 mx-auto'>
@@ -171,13 +171,13 @@
     </body>
     
     <script src = "https://code.jquery.com/jquery-3.6.0.min.js" integrity = "sha256-/xUj+3OJU5yExlq6GSYGSHk7tPXikynS7ogEvDej/m4=" crossorigin = "anonymous"></script>
-    <script src = "./js/jquery.rowspanizer.min.js"></script>
+    <script src = "/js/jquery.rowspanizer.min.js"></script>
     <link rel = "stylesheet" href = "//code.jquery.com/ui/1.13.0/themes/base/jquery-ui.css">
     <script src = "https://code.jquery.com/ui/1.13.0/jquery-ui.js"></script>
     <script src = "https://html2canvas.hertzen.com/dist/html2canvas.min.js"></script>
     <link id = "theme" href = "themes/hot-sneaks/jquery-ui.min.css" rel = "stylesheet">
-    <script type = "module" src = "./js/main.js"></script>
-    <script type = "module" src = "./js/init.js"></script>
-    <script type = "module" src = "./js/course.js"></script>
-    <script src = "./js/func.js"></script>
+    <script type = "module" src = "/js/main.js"></script>
+    <script type = "module" src = "/js/init.js"></script>
+    <script type = "module" src = "/js/course.js"></script>
+    <script src = "/js/func.js"></script>
 </html>
